feat: Create package of underwriting for division
