+++ src/main/java/command/UnderwritingCommand.java
@@ -1,8 +1,9 @@
 package command;
 
-import command.menu.UnderwritingMenu;
+import command.menu.sales.SalesManagement;
+import command.menu.underwriting.Underwriting;
+import command.menu.underwriting.UnderwritingManagement;
 import command.parser.UnderwritingParser;
-import domain.Underwriting;
 import java.util.List;
 import service.impl.UnderwritingServiceImpl;
 
@@ -11,19 +12,24 @@ public class UnderwritingCommand extends Command{
   private static final UnderwritingParser underwritingParser = UnderwritingParser.getInstance();
   private static final UnderwritingServiceImpl underwritingImpl = new UnderwritingServiceImpl();
   public static void run(){
-    printMenu("인수심사", UnderwritingMenu.values());
+    printMenu("인수심사", Underwriting.values());
   }
 
   public static void createAcceptancePolicy(){
     System.out.println("\n[인수정책 수립]");
-    List<Underwriting> underwritingList = underwritingImpl.createAcceptancePolicy();
+    List<domain.Underwriting> underwritingList = underwritingImpl.createAcceptancePolicy();
     printTable(underwritingList);
+
+    System.out.println("\n[인수정책을 수립해주세요]");
 
 
   }
 
   public static void underwrite(){
 
+
+
+    printMenu("인수심사", UnderwritingManagement.values());
   }
 
   public static void manageCollaboration(){
