+++ src/main/java/command/menu/underwriting/Underwriting.java
@@ -1,0 +1,28 @@
+package command.menu.underwriting;
+
+
+import command.UnderwritingCommand;
+import command.menu.Menu;
+
+public enum Underwriting implements Menu {
+  AcceptancePolicyCreation("인수정책 수립", UnderwritingCommand::createAcceptancePolicy),
+  LossRateManagement("손해율 관리", UnderwritingCommand::manageLossRate),
+  Underwriting("인수심사", UnderwritingCommand::underwrite);
+
+  private final String label;
+  private final Runnable runnable;
+
+  Underwriting(String label, Runnable runnable) {
+    this.label = label;
+    this.runnable = runnable;
+  }
+
+  public void execute() {
+    runnable.run();
+  }
+
+  @Override
+  public String toString() {
+    return this.label;
+  }
+}
